# [Merged by Bors] - meshdb: do not recompute ID and MinerID when block is loaded from database