## Motivation

blk.Initialize function is very slow, see the difference in benchmark:

```
benchmark                      old ns/op     new ns/op     delta
BenchmarkGetBlockHeader-16     119538        2810          -97.65%
```

It computes MinerID from signature and ID from encoded block every time when we load block from database.

related: https://github.com/spacemeshos/go-spacemesh/issues/2478

## Changes
Introduce DBBlock type that will encode ID and MinerID as a part of the block when block is saved to the local database.

## Test Plan
unit tests
## TODO
<!-- This section should be removed when all items are complete -->
- [x] Explain motivation or link existing issue(s)
- [x] Test changes and document test plan
- [x] Update documentation as needed

## DevOps Notes
<!-- Please uncheck these items as applicable to make DevOps aware of changes that may affect releases -->
- [x] This PR does not require configuration changes (e.g., environment variables, GitHub secrets, VM resources)
- [x] This PR does not affect public APIs
- [x] This PR does not rely on a new version of external services (PoET, elasticsearch, etc.)
- [x] This PR does not make changes to log messages (which monitoring infrastructure may rely on)