@dshulyak 2021-07-15
@antonlerner it appears that fetcher can actually request any data that is stored in the database in the raw format. how I can change the way block is stored in  the database without causing changes all over the place?