@dshulyak 2021-07-15
I made a wrapper that protects block store from internal changes. It still relies on a fact that the key will remain the same, but I think this is fine.