@noamnelke 2021-07-26
I know this is merged already and I think it was a good "quick fix", but I wonder if we can't implement for blocks the same mechanism we have for ATXs? @dshulyak 

The gist of it is that we always index ATXs by ID, so whenever we fetch an ATX from the DB we already know its ID - so we can just set it without calculating it. The only place we calculate an ATX ID is when receiving it over the network and then it's unavoidable (trust issue).

My question is - can we use the same principle for blocks (saving some space and an additional type + being more consistent) or is there some other scenario with blocks that makes this impossible? If it's not impossible - can you please create a tech-debt issue for this?