@dshulyak 2021-07-26
@noamnelke i noticed that we can do that. it is only 32b, and given that whole block was 10kb on a testnet saving space didn't seem important to me. i decided that it is better to have a future-proof id representation that won't change if index has to change. but overall it doesn't matter to me much, if you want i can remove it from DBBlock and recover from ID.